% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Breton.R
\name{theme_breton}
\alias{theme_breton}
\title{Theme Breton for ggplot2 Plots}
\usage{
theme_breton()
}
\value{
Applies Breton style to an existing ggplot2 plot
}
\description{
Applies Breton look and feel to an existing ggplot2 plot.
}
\examples{
mtcars |>
ggplot2::ggplot(ggplot2::aes(x = cyl, group = cyl, fill = as.factor(cyl))) +
ggplot2::geom_bar() +
ggplot2::labs(title = "Ggplot2 Plot",
     caption = "Nice plot") +
theme_breton()

}
